reefer-container-maintenance-predictor/requirements.txt
scikit-learn==0.19.0
pandas
psycopg2-binary
requests
watson_machine_learning_client
pymongo